Delegate Runner creation to a factory
